Updates to elicitation prompt layout
